+++ core/CHANGELOG.md
@@ -1,1 +1,5 @@
 ## Spree 2.2.0 (unreleased) ##
+
+* How Spree.user_class class attribute accepts a Symbol object.
+
+    *Bakhtiyor Hamidi*